review: bring back comments, symlink resolve, web target files consistency
