+++ engine/pill_web/Cargo.toml
@@ -1,0 +1,37 @@
+# pill_web is the web-target runtime, peer to pill_native. The crate is
+# empty on non-wasm builds (`#![cfg(target_arch = "wasm32")]` in lib.rs)
+# and all wasm-only deps are gated behind `[target.'cfg(target_arch = "wasm32")']`
+# so workspace-wide native builds compile pill_web as a no-op.
+#
+# Per-game wiring (the WebGame struct, embedded config.ini) still lives in
+# the launcher's wasm template (pill_web_app), which depends on this crate
+# and provides a thin `#[wasm_bindgen(start)]` shim that calls pill_web::run.
+
+[package]
+name = "pill_web"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["rlib"]
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+pill_engine = { path = "../pill_engine", features = ["game", "internal"] }
+pill_renderer = { path = "../pill_renderer" }
+pill_core = { path = "../pill_core" }
+
+wasm-bindgen = "0.2"
+wasm-bindgen-futures = "0.4"
+web-sys = { version = "0.3", features = [
+    "Window",
+    "Document",
+    "HtmlCanvasElement",
+    "Element",
+    "Performance",
+] }
+console_error_panic_hook = "0.1"
+console_log = "1"
+log = "0.4"
+winit = { version = "0.30.12" }
+config = "0.11"
+getrandom = { version = "0.3", features = ["wasm_js"] }
